requirements.txt
uagents
uagents-core
requests
nltk
python-dotenv
aiohttp
asyncio
tenacity
backoff
pandas
colorlog
Crewai
Pydantic
beautifulsoup4